DOMA-2041 fixed edge case with timestamps comparsion in test case for `syncSubscriptions`
A series of test cases was failing sometimes because of dates difference inspection was implemented in a way of comparing datetime representation of timestamps, cropped up to minute.
Replaced to inspecting a difference between timestamps in specified units.
In current cases we expect a difference in minutes to be zero, that means, that it not exceeds one minute.