DOMA-2236 fixed crash of Keystone while `changeTrackable` module processed fields of type `Select`
When we are trying to add a new field of type `Select` to Keystone `Ticket` schema (true for any schema), an util `makemigrations` will crash.
It crashes because we are picking not enough attributes from original field from `Ticket` to be able to programmatically create a corresponding field for `TicketChange` model (true for any model, that will hold trackable changes).