refactor(condo): DOMA-5608 function expectValuesOfCommonFields
In each test module a created schema object is expected to contain common set of fields. These checks spanning on multiple lines and gets repeated in almost every test module. Because of this duplication, they are presented not in every test module, but should.
Proposed test function makes expectation checks easier and removes a lot of duplicated code